+++ packages/react-together/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
 `0.3.0-2`
  - Added `resetOnDisconnect` option flag in `useStateTogether`
+ - Added `resetOnDisconnect` option flag in `useStateTogetherWithPerUserValues`
 `0.3.0-1`
  - Include `CHANGELOG.md` in package
 `0.3.0-0`